python-ocr/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
azure-cognitiveservices-vision-computervision==0.9.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
pydantic==2.4.2
pillow==10.1.0
numpy==1.26.2
python-jose==3.3.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1